Create LoadingWrapper.svelte component, move progressbar logic outside layout
